requirements_win.txt
json-repair
absl-py
accelerate
aiofiles
aiohappyeyeballs
aiohttp
aiosignal
altair
annotated-types
anyascii
anyio
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
astunparse
async-lru
attrs
babel
beautifulsoup4
bert-score
bleach
BLEURT @ git+https://github.com/google-research/bleurt.git@cebe7e6f996b40910cfaa520a63db47807e3bf5c
blis
blobfile
cachetools
catalogue
certifi
cffi
charset-normalizer
click
cloudpathlib
colorama
comm
confection
contourpy
contractions
cycler
cymem
cytoolz
datasets
debugpy
decorator
defusedxml
dill
distro
emoji
en_core_web_md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.8.0/en_core_web_md-3.8.0-py3-none-any.whl#sha256
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256
evaluate
executing
Faker
fastapi
fastjsonschema
ffmpy
filelock
fire
flatbuffers
floret
fonttools
fqdn
frozenlist
fsspec
ftfy
gast
git-filter-repo
google-pasta
gradio
gradio_client
graphviz
greenlet
grpcio
h11
h5py
httpcore
httpx
huggingface-hub
idna
importlib_resources
ipykernel
ipython
ipython-genutils
ipython-sql
isoduration
jedi
jellyfish
jieba
Jinja2
jiter
joblib
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
keras
kiwisolver
langcodes
language_data
libclang
llama_models
llama_stack
llama_stack_client
lxml
marisa-trie
Markdown
markdown-it-py
MarkupSafe
matplotlib
matplotlib-inline
mdurl
mistune
ml_dtypes
mpmath
multidict
multiprocess
murmurhash
mysql-connector-python
namex
narwhals
nbclient
nbconvert
nbformat
nest-asyncio
networkx
nltk
notebook
notebook_shim
numpy
openai
opt_einsum
optree
orjson
overrides
packaging
pandas
pandocfilters
parso
peft
pexpect
pillow
platformdirs
portalocker
preshed
prettytable
prometheus_client
prompt_toolkit
propcache
protobuf
psutil
ptyprocess
pure_eval
pyahocorasick
pyaml
pyarrow
pycparser
pycryptodomex
pydantic
pydantic_core
pydub
Pygments
pyparsing
pyphen
python-dateutil
python-dotenv
python-json-logger
python-multipart
pytz
PyYAML
pyzmq
RapidFuzz
referencing
regex
requests
rfc3339-validator
rfc3986-validator
rich
rouge-chinese
rouge_score
rpds-py
sacrebleu
safetensors
scikit-learn
scipy
semantic-version
Send2Trash
sentence-transformers
sentencepiece
setuptools
shellingham
six
smart-open
sniffio
soupsieve
spacy
spacy-legacy
spacy-loggers
SQLAlchemy
sqlparse
srsly
sse-starlette
stack-data
stanfordnlp
stanza
starlette
sympy
tabulate
tensorboard
tensorboard-data-server
tensorflow
termcolor
terminado
textacy
textsearch
tf-slim
tf_keras
thinc
threadpoolctl
tiktoken
tinycss2
tokenizers
toolz
torch
tornado
tqdm
traitlets
transformers
trl
typer
types-python-dateutil
typing_extensions
tzdata
uri-template
urllib3
uvicorn
wasabi
wcwidth
weasel
webcolors
webencodings
websocket-client
websockets
Werkzeug
wheel
wrapt
xxhash
yarl
